Fix for issue 42:
Return null from DefaultSecurityConfiguration#getResourceFile(String)
when the argument is null. Update the java doc to include null as a
return meaning the resource was not found. This was already the case if
the file was not found but is also the case if the file name is null.

This allows the SecurityConfigurationTest to pass.