Fix to IntegerAccessReferenceMapTest test failures.
These failures were caused by a use before initialization bug in the
counter used for IntegerAccessReferenceMap unique references. This
counter is a instance field. The constructors that take Sets would
call the AbstractAccessReferenceMap constructor with that set. The
super constructor would then call the abstract getUniqueReference()
implemented in IntegerAccessReferenceMap. getUniqueReference()
in IntegerAccessReferenceMap would then use the counter field that
had not been initalized. Once the call to super(Set) was finished,
IntegerAccessReferenceMap's constructors would continue and finally
initialize the counter to 1. If the set passed had more than one entry
duplicate ids could occur. Additionally, the reference id of 0 would be
used despite the counter being initilized to 1. We are lucky that the
unit test accidentally triggered this behavior.

Though simply not initializing the counter to 1 and letting it get the
default 0 does fix the test, the issue is not clear and could still be
a problem in other implementations.

Instead, AbstractAccessReferenceMap(Set)
and AbstractAccessReferenceMap(Set,int) have been
deprecated. RandomAccessReferenceMap and IntegerAccessReferenceMap
have been changed to not call these constructors and call update(Set)
themselves after they initialize.

There is still the possibility that a third party subclassing of
AbstractAccessReferenceMap or IntegerAccessReferenceMap that implement
getUniqueReference() could trigger the same type of bug. It might be
worth considering making getUniqueReference() in these classes final.

Additionally, I have enabled showWarnings and showDeprecation for
compilation in the pom so that such deprecations will show up. This
produces considerable noise currently as there are other deprecated apis
in use which should probably be fixed.